change hatch command to extract in newdir
The hatch command extracts archives on the current working directory
in most cases this is not what users want as this command is
most likely used on the root of the current engines directory.

Theses changes modify the hatch command so that extraction is done
inside a newly created directory named after the repository or
tarball used. If the directory already exists however, it's
contents will be overwritten.

* add handling for lowercase and uppercase confirmation