add subenv.yml aliases handling (#29)
The `subenv.yml` is a useful file for storing subenv specific
settings that you might need for your project. one common usecase
for this file would be to integrate custom aliases specific to your
project. However, this feature wasn't supported.

These changes now make it possible to add an aliases section to
your `subenv.yml` the same way as you would in your default
`engine.yml`.

This is done by navigating to the specific sub-envs path and
loading up the `subenv.yml` inside the engine and merging
both aliases sections when aliases are accessed.